Added migrations in order to shift models to invite app
The InviteFriendsPage and the FriendInvitation were added to the
invite app and removed from the home app. The corresponding
migrations were created for the same.
