+++ project/invite/templates/invite/thankyou.html
@@ -1,0 +1,7 @@
+{% extends "base.html" %}
+{% load wagtailcore_tags %}
+
+{% block content %}
+    <h1> Thanks for inviting!</h1>
+    <p><a href="{{ page.get_parent.url }}">Return to home</a></p>
+{% endblock %}
